# filmbox
In today's world, people have become addicted to their phones. Phones are everything to
humans. People can use it for anything and that's where the idea for the Filmbox app comes
from. There is a huge archive of movies on the Internet. It gets harder and harder to choose
from some romantic and some drama. People keep looking for an app where they can watch,
search for movies, see their rating and description, and see what their friends have watched
In the Filmbox application, the user will be able to record the movies they watch and see
how many movies they have watched in total, and they will be able to review the movies watched by other users.

In this project the developer used tmdb database as api source and firebase authentication, firestore and real
time database.